Changed everything
During my vacation at Madeira I decided to realise that project with a
new logical background. So I changed the concept as known as the
ClassDiagramm and implemented the new logic.
The ClassDiagram is now on not up to date meaning that some methods are
missing there. This will get changed in some days ;)
I also decided to create a brand new graphical interface because the
current one looks like Windows 95… :/
I think the vacation were needed to come away from the stress at school
and to earn new energy for realising that project :)